refactor(interp): simplify jit plan graph
