+++ interp/trace_test.go
@@ -321,6 +321,24 @@ func TestTracer_Headers(t *testing.T) {
 	})
 }
 
+func TestTracer_IsolatesPrograms(t *testing.T) {
+	tracer := NewTracer()
+	first := program.New([]instr.Instruction{instr.New(instr.I32_CONST, 1)})
+	second := program.New([]instr.Instruction{instr.New(instr.I32_CONST, 2)})
+
+	left := New(first, WithTracer(tracer), WithThreshold(-1))
+	defer left.Close()
+	before := tracer.codes(left)
+
+	right := New(second, WithTracer(tracer), WithThreshold(-1))
+	defer right.Close()
+	after := right.tracer.codes(right)
+
+	require.Same(t, tracer, left.tracer)
+	require.NotSame(t, tracer, right.tracer)
+	require.NotSame(t, &before[0][0], &after[0][0])
+}
+
 func TestTracer_Remove(t *testing.T) {
 	tracer := NewTracer()
 	first := program.New([]instr.Instruction{
